# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreatePaymentRequest:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      currency:
        type: string
      description:
        maxLength: 500
        type: string
      gateway:
        enum:
        - stripe
        - omise
        - paypal
        - manual
        type: string
      invoice_id:
        type: string
      method:
        enum:
        - credit_card
        - bank_transfer
        - cash
        - qr_code
        type: string
    required:
    - amount
    - booking_id
    - currency
    - method
    type: object
  dto.CreateRoomRequest:
    properties:
      amenities:
        items:
          type: string
        type: array
      deposit:
        minimum: 0
        type: number
      description:
        maxLength: 2000
        type: string
      floor:
        maximum: 200
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
      number:
        maxLength: 20
        minLength: 1
        type: string
      photos:
        items:
          type: string
        type: array
      rent_price:
        minimum: 0
        type: number
      size_sqm:
        minimum: 0
        type: number
      type:
        enum:
        - studio
        - 1br
        - 2br
        - suite
        - loft
        type: string
    required:
    - name
    - number
    - rent_price
    - type
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
...
  dto.PaymentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PaymentResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.PaymentResponse:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      gateway:
        type: string
      gateway_tx_id:
        type: string
      id:
        type: string
      invoice_id:
        type: string
      method:
        type: string
      owner_id:
        type: string
      paid_at:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RoomListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.RoomResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.RoomResponse:
    properties:
      amenities:
        items:
          type: string
        type: array
      created_at:
        type: string
      deposit:
        type: number
      description:
        type: string
      floor:
        type: integer
      id:
        type: string
      name:
        type: string
      number:
        type: string
      owner_id:
        type: string
      photos:
        items:
          type: string
        type: array
      rent_price:
        type: number
      size_sqm:
        type: number
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateRoomRequest:
    properties:
      amenities:
        items:
          type: string
        type: array
      deposit:
        minimum: 0
        type: number
      description:
        maxLength: 2000
        type: string
      floor:
        maximum: 200
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
      photos:
        items:
          type: string
        type: array
      rent_price:
        minimum: 0
        type: number
      size_sqm:
        minimum: 0
        type: number
      status:
        enum:
        - available
        - occupied
        - maintenance
        type: string
      type:
        enum:
        - studio
        - 1br
        - 2br
        - suite
        - loft
        type: string
    type: object
...
  inbound.IncomeExpenseReport:
    properties:
      invoiceCount:
        format: int64
        type: integer
      month:
        type: integer
      outstanding:
        format: float64
        type: number
      ownerID:
        type: string
      paymentCount:
        format: int64
        type: integer
      totalIncome:
        format: float64
        type: number
      totalInvoiced:
        format: float64
        type: number
      year:
        type: integer
    type: object
  inbound.OwnerDashboard:
    properties:
      activeBookings:
        format: int64
        type: integer
      availableRooms:
        format: int64
        type: integer
      maintenanceRooms:
        format: int64
        type: integer
      monthlyRevenue:
        format: float64
        type: number
      occupiedRooms:
        format: int64
        type: integer
      overdueInvoices:
        format: int64
        type: integer
      ownerID:
        type: string
      pendingBookings:
        format: int64
        type: integer
      pendingMaintenance:
        format: int64
        type: integer
      totalRooms:
        format: int64
        type: integer
      unreadMessages:
        format: int64
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: Production-ready Go REST API starter with hexagonal architecture.
  title: Go Starter API
  version: "1.0"
paths:
...
  /api/v1/owners/{id}/dashboard:
    get:
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inbound.OwnerDashboard'
      security:
      - BearerAuth: []
      summary: Owner dashboard summary
      tags:
      - reports
  /api/v1/payments:
    get:
      parameters:
      - description: Filter by booking
        in: query
        name: booking_id
        type: string
      - description: Filter by tenant
        in: query
        name: tenant_id
        type: string
      - description: Filter by owner
        in: query
        name: owner_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentListResponse'
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Create a payment record
      tags:
      - payments
  /api/v1/payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Get payment by ID
      tags:
      - payments
  /api/v1/payments/{id}/refund:
    post:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Refund a payment
      tags:
      - payments
  /api/v1/payments/webhook/{gateway}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Gateway name (stripe|omise|paypal)
        in: path
        name: gateway
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Payment gateway webhook
      tags:
      - payments
  /api/v1/reports/income:
    get:
      parameters:
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inbound.IncomeExpenseReport'
      security:
      - BearerAuth: []
      summary: Income and expense report
      tags:
      - reports
  /api/v1/rooms:
    get:
      parameters:
      - description: Filter by owner ID
        in: query
        name: owner_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Min rent price
        in: query
        name: min_price
        type: number
      - description: Max rent price
        in: query
        name: max_price
        type: number
      - description: Search by name/number
        in: query
        name: search
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RoomListResponse'
      summary: List rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      parameters:
      - description: Room data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a room
      tags:
      - rooms
  /api/v1/rooms/{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a room
      tags:
      - rooms
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RoomResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - rooms
    put:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RoomResponse'
      security:
      - BearerAuth: []
      summary: Update a room
      tags:
      - rooms
  /api/v1/tenants/{id}/history:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inbound.IncomeExpenseReport'
      security:
      - BearerAuth: []
      summary: Tenant payment history
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"